+++ index.html
@@ -7,8 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
 </head>
 
 <body>
@@ -228,6 +226,8 @@
       </div>
     </section>
   </template>
+
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
