+++ js/data.js
@@ -1,0 +1,37 @@
+const NAMES = [
+  'Александр',
+  'Михаил',
+  'Артем',
+  'Марк',
+  'Дмитрий',
+  'Мария',
+  'Ева',
+  'Варвара',
+  'Виктория',
+  'Александра',
+];
+const DESCRIPTIONS = [
+  'Закат на пляже',
+  'Город в тумане',
+  'Пейзаж в горах',
+  'Праздник в городе',
+  'Семья на отдыхе',
+  'Животные в зоопарке',
+  'Архитектура города',
+  'Природа вблизи',
+  'Спортсмены на стадионе',
+  'Музыканты на концерте',
+];
+const MESSAGE = [
+  "Всё отлично!",
+  "В целом всё неплохо. Но не всё.",
+  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!",
+];
+
+const PHOTO_DESCRIPTION_COUNTER = 25;
+
+export {NAMES, DESCRIPTIONS, MESSAGE, PHOTO_DESCRIPTION_COUNTER};
+
