+++ js/util.js
@@ -1,3 +1,16 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+}
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, getRandomArrayElement};
+
+
 function checkStringLength(len, maxLength){
   return len >= maxLength;
 }
@@ -20,3 +33,6 @@ function digitsToInteger(input) {
  const numberResult = parseInt(result);
  return result.length > 0 ? numberResult : NaN;
 }
+
+export {checkStringLength, isPalindrome, digitsToInteger};
+
